requirements.txt
PyYAML==6.0.3
numpy==2.4.2
pandas==3.0.0
openpyxl==3.1.5
